requirements.txt
idna==3.4
joblib==1.2.0
numpy==1.24.3
pandas==2.0.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
Pillow==9.3.0
requests==2.28.1
scikit-learn==1.3.1
scipy==1.9.3
sklearn==0.0.post1
threadpoolctl==3.1.0
torch==2.0.1
torchvision==0.15.2
umap-learn==0.5.5
matplotlib==3.7.2
seaborn==0.12.2